requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bokeh==3.7.3
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.3
cramjam==2.11.0
curl_cffi==0.12.0
dask==2025.7.0
dataclasses-json==0.6.7
dateparser==1.2.2
debugpy==1.8.14
decorator==4.4.2
defusedxml==0.7.1
distributed==2025.7.0
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
executing==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastparquet==2024.11.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.73.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-groq==0.3.6
langchain-text-splitters==0.3.9
langgraph==0.6.2
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.2
langgraph-sdk==0.2.0
langsmith==0.4.8
lightgbm==4.6.0
llvmlite==0.44.0
locket==1.0.0
lxml==6.0.0
lz4==4.4.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.92.2
opencv-python==4.12.0.88
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
optuna==4.4.0
orjson==3.11.0
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
partd==1.4.2
peewee==3.18.2
phidata==2.7.10
pillow==11.3.0
platformdirs==4.3.8
playwright==1.53.0
posthog==5.4.0
primp==0.15.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shap==0.48.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
streamlit==1.46.1
sympy==1.14.0
tblib==3.1.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.20.1
yfinance==0.2.65
youtube-transcript-api==1.2.1
zict==3.0.0
zipp==3.23.0
zstandard==0.23.0
